Move QnCheckBoxedHeaderView to a separate file
